+++ account-keeper-sdk/src/main/java/com/dwarfeng/acckeeper/sdk/bean/BeanMapper.java
@@ -281,4 +281,14 @@ public interface BeanMapper {
 
     @InheritInverseConfiguration
     StaticLoginInfo staticLoginInfoFromWebInput(WebInputStaticLoginInfo webInputStaticLoginInfo);
+
+    FastJsonLoginStateLookupResult loginStateLookupResultToFastJson(LoginStateLookupResult loginStateLookupResult);
+
+    @InheritInverseConfiguration
+    LoginStateLookupResult loginStateLookupResultFromFastJson(FastJsonLoginStateLookupResult fastJsonLoginStateLookupResult);
+
+    WebInputLoginStateLookupInfo loginStateLookupInfoToWebInput(LoginStateLookupInfo loginStateLookupInfo);
+
+    @InheritInverseConfiguration
+    LoginStateLookupInfo loginStateLookupInfoFromWebInput(WebInputLoginStateLookupInfo webInputLoginStateLookupInfo);
 }